bgpd: last_reset_cause_size can't be larger than 65k bytes
Reduce the scope, to avoid comparing uint16_t vs. size_t in a loop.

```
vty_out(vty,
	"  Message received that caused BGP to send a NOTIFICATION:\n    ");
	for (i = 1; i <= p->last_reset_cause_size;
	     i++) {
```

+++ bgpd/bgpd.h
@@ -1637,7 +1637,7 @@ struct peer {
 	 * a new value to the last_reset reason
 	 */
 
-	size_t last_reset_cause_size;
+	uint16_t last_reset_cause_size;
 	uint8_t last_reset_cause[BGP_MAX_PACKET_SIZE];
 
 	/* The kind of route-map Flags.*/
